Rework sandbox create flow and refactor build sandbox, adjust prompt ordering
