+++ packages/cli-lib/lang/en.lyaml
@@ -653,6 +653,13 @@ en:
                 describe: "Name to use for created sandbox"
               type:
                 describe: "Type of sandbox to create (standard | development)"
+            failure:
+              optionMissing:
+                type: "Invalid or missing sandbox --type option in command. Please try again."
+                name: "Invalid or missing sandbox --name option in command. Please try again."
+              creatingWithinSandbox: "Sandboxes must be created from a production account. Your current default account {{#bold}}{{ sandboxName }}{{/bold}} is a {{ sandboxType }} sandbox.
+                \n- Run {{#bold}}hs accounts use{{/bold}} to switch to your default account to your production account.
+                \n- Run {{#bold}}hs auth{{/bold}} to connect a production account to the HubSpot CLI.\n"
           delete:
             describe: "Delete a sandbox account"
             debug:
@@ -1053,12 +1060,6 @@ en:
           success:
             configFileUpdated: "{{ configFilename }} updated with {{ authMethod }} for account {{ account }}."
           failure:
-            optionMissing:
-              type: "Invalid or missing sandbox --type option in command. Please try again."
-              name: "Invalid or missing sandbox --name option in command. Please try again."
-            creatingWithinSandbox: "Sandboxes must be created from a production account. Your current default account {{#bold}}{{ sandboxName }}{{/bold}} is a {{ sandboxType }} sandbox.
-              \n- Run {{#bold}}hs accounts use{{/bold}} to switch to your default account to your production account.
-              \n- Run {{#bold}}hs auth{{/bold}} to connect a production account to the HubSpot CLI.\n"
             limit:
               developer:
                 one: "{{#bold}}{{ accountName }}{{/bold}} reached the limit of {{ limit }} development sandbox per account. View sandbox details at {{ link }}. To connect a sandbox to your HubSpot CLI, run {{#bold}}hs auth{{/bold}} and follow the prompts."
@@ -1095,8 +1096,8 @@ en:
               standard: "Sync all supported assets to {{#cyan}}{{#bold}}{{ sandboxName }}{{/bold}}{{/cyan}} from {{#bold}}{{ parentAccountName }}{{/bold}}?"
               developer: "Sync CRM object definitions to {{#cyan}}{{#bold}}{{ sandboxName }}{{/bold}}{{/cyan}} from {{#bold}}{{ parentAccountName }}{{/bold}}?"
             syncContactRecords:
-              standard: "Do you want the sync to include up to 5000 contact records and associations? Syncing may take some time"
-              developer: "Do you want the sync to include up to 100 contact records and associations?"
+              standard: "Include up to 5000 contact records and associations in this sync? Syncing contact records may take some time"
+              developer: "Include up to 100 contact records and associations in this sync? Syncing contact records may take some time"
           loading:
             startSync: "Initiating sync..."
             fail: "Failed to sync sandbox."
